Use puppetlabs-apt for APT Sources
Prior to this commit, the boundary module created its own resources
for managing Boundary's APT sources.

This commit replaces those resources with a single apt::source
resource along with two pieces of logic to properly craft the APT
source based on whether the OS is Debian vs Ubuntu.

This work creates a dependency on puppetlabs-apt and through that
module, lsb-release (for access to $::lsbdistcodename).